% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ai_asso_processed}
\alias{ai_asso_processed}
\title{AI association data}
\format{
The function returns the \code{associatoR} object, which contains a list with the following elements, each a \code{tibble}:
\describe{
 \item{participant}{Index of the element in \code{text} where match was found. Formatted as a factor with the number of levels matching the original number of documents.}
 \item{cues}{Label of the SDG found in document.}
 \item{responses}{The name of the ensemble system that produced the match.}
 \item{hit}{Index of hit for the Ensemble model.}
}
}
\source{
\url{https://ap-unsdsn.org/regional-initiatives/universities-sdgs/}
}
\usage{
ai_asso_processed
}
\description{
\code{ar_import} processes R data and generates an associatoR object
}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv
}
\keyword{datasets}
