% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_characterize_targets}
\alias{ar_characterize_targets}
\title{Characterize targets using external lookup-table}
\usage{
ar_characterize_targets(
  associations,
  characteristic = c("valence", "arrousal", "dominance", "concreteness", "frequency")
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{characteristic}{a \code{character} string or vector specifying which word characteristics to add to \code{targets}. Must be one or multiple of \code{c("valence", "frequency", "pos", "concreteness")}.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles, with \code{targets} gaining additional columns including the target characteristics:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
 \item{targets}{A tibble of targets including the specified analysis target, and word characteristics}
}
}
\description{
Append target characteristic to the \code{targets} table in the \code{associatoR} object
}
\references{
Warriner, A.B., Kuperman, V. & Brysbaert, M. Norms of valence, arousal, and dominance for 13,915 English lemmas. Behav Res 45, 1191–1207 (2013). https://doi.org/10.3758/s13428-012-0314-x
  Brysbaert, M., Warriner, A.B. & Kuperman, V. Concreteness ratings for 40 thousand generally known English word lemmas. Behav Res 46, 904–911 (2014). https://doi.org/10.3758/s13428-013-0403-5
}
