% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{ar_cluster_stability}
\alias{ar_cluster_stability}
\title{Cluster stability}
\usage{
ar_cluster_stability(associations, n_boot = 1000, unique = FALSE)
}
\arguments{
\item{associations}{an \code{associatoR} object including target_embeddings.}

\item{n_boot}{an \code{integer} specifying the number of bootstrap samples to draw. Default is \code{1000}.}

\item{unique}{a \code{logical} specifying whether duplicate targets in bootstrap shouod be ignored. Default is \code{TRUE}.}
}
\value{
The function returns a list containing two matrices with the target- (\code{target_stability}) and cluster-wise (\code{cluster_stability}) probabilities of being assigned to the same cluster under random pertubation.
}
\description{
\code{ar_cluster_stability} evaluates the stability of a clustering generated by ar_cluster_targets()
}
\details{
Clustering stability is calculated with the identical clustering settings (e.g., clustering algorithm, similarity function, etc.) used to generate the original cluster assignment. The information is stored as a attribute to the \code{targets} table with name \code{"cluster_settings"}.
}
\examples{
ar_obj = ar_import(intelligence,
                   participant = participant_id,
                   cue = cue,
                   response = response,
                   participant_vars = c(gender, education),
                   response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_embed_targets() \%>\%
  ar_cluster_targets(method = "louvain")

ar_cluster_stability(ar_obj)

}
