% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{ar_cluster_targets}
\alias{ar_cluster_targets}
\title{Cluster targets}
\usage{
ar_cluster_targets(associations, method = c("louvain"), ...)
}
\arguments{
\item{associations}{an \code{associatoR} object including target_embeddings.}

\item{method}{a \code{character} specifying the clustering method. One of \code{c("louvain")}. Default is \code{"louvain"}.}

\item{...}{additional attributes passed on to the clustering function.}
}
\value{
The function returns the \code{associatoR} object including a new
  \code{clusters} column in \code{targets} containing the target clusters.
}
\description{
Cluster targets
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_embed_targets() \%>\%
  ar_cluster_targets(method = "louvain")

}
