% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{ar_compare_embeddings}
\alias{ar_compare_embeddings}
\title{Embedding similarity comparison}
\usage{
ar_compare_embeddings(
  associations,
  participant_vars,
  type = "triangle",
  intersection = "pair",
  ...
)
}
\arguments{
\item{associations}{an \code{associatoR} object including target_embeddings.}

\item{participant_vars}{one or more column names specifying the grouping variables for embedding comparisons.}

\item{type}{a \code{character} specifying whether to compute representational similarity based on the full \code{"triangle"} or \code{"row"}-wise. Default is \code{"triangle"}.}

\item{intersection}{a \code{character} specifying whether to compute representations for the set of targets shared by all groups (\code{"all"}) or only by the individual pair (\code{"pair"}). The default is (\code{"pair"}).}

\item{...}{arguments passed on to \code{ar_embed_targets}. If no arguments are specified arguments are taken from an existing \code{target_embedding} or based on default values.}
}
\value{
The function returns a table of representation similarities.
}
\description{
\code{ar_compare_embeddings} compares embeddings between groups using representational similarity analysis
}
\details{
Representational similarity is calculated based on the Spearman correlation between cosine similarity matrices extracted from each embedding specified by one or more grouping factors.
}
\examples{
ar_obj = ar_import(intelligence,
                   participant = participant_id,
                   cue = cue,
                   response = response,
                   participant_vars = c(gender, education),
                   response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_embed_targets()

ar_compare_embeddings(ar_obj, gender)

}
