% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{ar_compare_targets}
\alias{ar_compare_targets}
\title{Compare target characteristics between participant groups}
\usage{
ar_compare_targets(associations, participant_vars, target_var, fun = mean, ...)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import} with targets defined by \link[associatoR]{ar_set_targets}.}

\item{participant_vars}{a variable name vector specifying the grouping variables.}

\item{target_var}{a variable name specifying the target variable to be summarized.}

\item{fun}{a \code{function} specifying the statistic to calculate from \code{target_var}, defaults to mean.}

\item{...}{additional arguments passed on to fun}
}
\value{
Returns a \link[tibble]{tibble} containing grouped summaries based on \code{fun}.
}
\description{
\code{ar_compare_targets} compares target characteristics between participant groups taking into account target frequencies in responses to return summary statistics.
}
\examples{
# calculate means by using fun = mean and passing na.rm argument to mean()
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_summarize_targets(response_var = response_position) \%>\%
  ar_compare_targets(participant_vars = c(gender, education), target_var = response_position, fun = mean, na.rm = TRUE)

}
