% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{ar_cross_targets}
\alias{ar_cross_targets}
\title{Create participant-targets cross tables}
\usage{
ar_cross_targets(associations, participant_vars, target_var, normalize = FALSE)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import} with targets defined by \link[associatoR]{ar_set_targets}.}

\item{participant_vars}{a variable name vector specifying the participant varables to group counts with.}

\item{target_var}{a variable name specifying the targets variable group counts with.}

\item{normalize}{a \code{logical} scalar, indicating if counts should be normalized to the participant varable groups. Defaults to \code{FALSE}.}
}
\value{
Returns a \link[tibble]{tibble} containing grouped counts, or grouped counts normalized within participant variable groups if \code{normalize = TRUE}.
}
\description{
\code{ar_cross_targets} creates tables counting targets among responses, for each combination of participant varable groups seperately.
}
\examples{
# one participant variable, no normalization
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_embed_targets() \%>\%
  ar_cluster_targets() \%>\%
  ar_cross_targets(participant_vars = gender, target_var = cluster)

# two participant variables, normalized
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_embed_targets() \%>\%
  ar_cluster_targets() \%>\%
  ar_cross_targets(participant_vars = c(gender, education), target_var = cluster, normalize = TRUE)
}
