% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_define_target}
\alias{ar_define_target}
\title{Define target in an associatoR object}
\usage{
ar_define_target(x, source = c("cues", "responses"), response_level = 1)
}
\arguments{
\item{x}{an \code{associatoR} object generated by associatoR::ar_import()}

\item{source}{one of \code{c("cues", "responses")} specifying, which column in $participants contains targets.}

\item{response_level}{an \code{integer} (\code{1}) specifying, which response-level contains targets.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{p_id} to identify participants, and potentially additional participant attributes \code{p_attr_*}.}
 \item{cues}{A tibble of all cues \code{c_cue} and potentially additional cue attributes \code{c_attr_*}.}
 \item{responses}{A tibble of all participant's \code{p_id} cue \code{r_cue} and response \code{r_response} pairs including response levels \code{r_level}, as well as potentially additional responseattributes \code{r_attr_*}.}
 \item{targets}{A tibble of \code{target}s that will be analyzed by subsequent functions.}
}
}
\description{
\code{ar_define_target} defines the target level to further analyze in an associatoR object. This could be multiple cue words to analyze based on responses, response words to analyze based on an external embedding, or response words to analyze based on a second level of responses.
}
\examples{

ai_asso <- ar_define_target(ai_asso_imported, source = "responses")

}
