% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ar_export_all}
\alias{ar_export_all}
\title{Join all associatoR tibbles into one}
\usage{
ar_export_all(associations)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}
}
\value{
a \code{tibble} containing all joined data in associations.
}
\description{
Join all associatoR tibbles into one
}
