% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import}
\alias{ar_import}
\title{Import association data}
\usage{
ar_import(
  data,
  participant,
  cue = NULL,
  cue_manual = NULL,
  response,
  participant_vars = NULL,
  cue_vars = NULL,
  response_vars = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble} containing the association data.}

\item{participant}{a \code{character} string of the variable name in \code{data} identifying participants.}

\item{cue}{a \code{character} string of the variable name in \code{data} identifying the cues.}

\item{cue_manual}{an optional \code{character} string specifying the cue for all responses. Overrides argument \code{cue}.}

\item{response}{a \code{character} string of the variable name in \code{data} identifying responses.}

\item{participant_vars}{an optional \code{character} vector containing all additional variable names in \code{data} that are attributes of participants.}

\item{cue_vars}{an optional \code{character} vector containing all additional variable names in \code{data} that are attributes of cues.}

\item{response_vars}{an optional \code{character} vector containing all additional variable names in \code{data} that are attributes of responses}

\item{verbose}{an \code{logical} specifying whether to show messages.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
}
}
\description{
\code{ar_import} processes tabular association data into an associatoR object.
}
\examples{
ai_asso_imported <- ar_import(ai_asso,
                              participant = "id",
                              response = "association_correct",
                              cue_manual = "AI",
                              participant_vars = c("age", "gender", "use", "expertise"),
                              response_vars = c("association", "trial"))

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
