% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import_multiple_cues_single_response_level}
\alias{ar_import_multiple_cues_single_response_level}
\title{ar_import_multiple_cues_single_response_level}
\usage{
ar_import_multiple_cues_single_response_level(
  data,
  cue,
  cue_attributes = c(),
  participant_id,
  participant_attributes = c(),
  response,
  response_attributes = c()
)
}
\arguments{
\item{response_attributes}{}
}
\description{
ar_import_multiple_cues_single_response_level
}
\examples{

ar_import_multiple_cues_single_response_level(data = fa_data,
                                              cue = "cue",
                                              cue_attributes = c(),
                                              participant_id = "participantID",
                                              participant_attributes = c("age", "gender"),
                                              response = "response",
                                              response_attributes = c("pos", "created_at"))

}
