% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import_single_cue_multiple_response_levels}
\alias{ar_import_single_cue_multiple_response_levels}
\title{ar_import_single_cue_multiple_response_levels}
\usage{
ar_import_single_cue_multiple_response_levels(
  data,
  cue,
  cue_attributes = c(),
  participant_id,
  participant_attributes = c(),
  response,
  response_level,
  response_attributes = c()
)
}
\arguments{
\item{response_attributes}{}
}
\description{
ar_import_single_cue_multiple_response_levels
}
\examples{

ar_import_single_cue_multiple_response_levels(data = risk_net,
                                              cue = "from_correct",
                                              cue_attributes = c(),
                                              participant_id = "subject_no",
                                              participant_attributes = c("gender", "age"),
                                              response = "to_correct",
                                              response_level = "to_correct_level",
                                              response_attributes = c("trial"))


}
