% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import_single_cue_single_response_level}
\alias{ar_import_single_cue_single_response_level}
\title{Title}
\usage{
ar_import_single_cue_single_response_level(
  data,
  cue,
  participant_id,
  participant_attributes,
  response,
  response_attributes
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble} containing the association data.}

\item{cue}{a \code{character} scalar to specify a cue that was used for all responses.}

\item{participant_id}{a \code{character} scalar of the variable name in \code{data} identifying individual participants.}

\item{participant_attributes}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of participants.}

\item{response}{a \code{character} scalar of the variable name in \code{data} identifying individual responses.}

\item{response_attributes}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of responses.}
}
\description{
Title
}
\examples{

ar_import_single_cue_single_response_level(data = ai_asso, cue = "AI",
                                           participant_id = "id",
                                           participant_attributes = c("age", "gender", "use", "expertise"),
                                           response = "association_correct",
                                           response_attributes = c("association", "trial"))

}
