% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize}
\alias{ar_normalize}
\title{Normalize responses in an associatoR object}
\usage{
ar_normalize(x, case = "lower", punct = "end", whitespace = "squish")
}
\arguments{
\item{x}{an \code{associatoR} object generated by \code{associatoR::ar_import()}}

\item{case}{one of \code{c("lower", "upper", "original")}, converts the resp_text to the specified casing, \code{"original"} does not change the casing.}

\item{punct}{one of \code{c("end", "all", "original")}, \code{end} replaces all punctuation at the end (or only followed by whitespace) of the string with whitespace, \code{"all"} replaces all punctuation characters in the string by whitespace, \code{"original"} changes nothing.}

\item{whitespace}{one of \code{c("squish", "trim", "original")}, \code{"squish"} removes whitespace from the ends of the string, an reduces multiple whitespace between characters to one, \code{"trim"} removes whitespace from the ends of the string, \code{"original"} keeps whitespaces as is (note that additional whitespaces can be generated based on the string passed to the \code{punct} argument).}
}
\value{
Returns an \code{associatoR} object containing a list of \code{tibble}s with \code{fun} applied to all responses:
\describe{
 \item{participants}{A tibble of all participants including \code{part_id}, and potentially additional participant-specific attributes \code{part_*}.}
 \item{cues}{A tibble of all cues including \code{cue_id}, \code{cue_text} and potentially additional cue-specific attributes \code{cue_*}.}
 \item{responses}{A tibble of all responses including \code{resp_id}, \code{resp_text} and potentially additional response-specific attributes \code{resp_*}.}
}
}
\description{
\code{ar_normalize} offers a set of often used normalization steps with sensible defaults and wraps around \code{ar_normalize_manual}.
}
\examples{

ai_asso_processed <- ar_normalize(ai_asso_processed)
ai_asso_processed <- ar_normalize(ai_asso_processed, case = "original",
                                  punct = "all")

}
