% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize_dict}
\alias{ar_normalize_dict}
\title{Normalize responses using dictionary}
\usage{
ar_normalize_dict(associations, dictionary, process_cues = FALSE)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{dictionary}{a \code{data.frame} consisting of two columns named \code{c("old","new")}.}

\item{process_cues}{a \code{logical} indicating if cues should be processed (i.e. changed by \code{fun}) or not.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:

\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes. All responses are passed through \code{fun}}
}
}
\description{
\code{ar_normalize_dict} offers an interface to normalization responses (and cues) in \code{associatoR} objects based on a dictionary.
}
\examples{
dict <- data.frame(old = c("intelligence", "iq"),
                   new = c("Intelligence", "IQ"))

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_normalize_dict(dict)

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
