% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize_manual}
\alias{ar_normalize_manual}
\title{Normalize words in an associatoR object manually}
\usage{
ar_normalize_manual(associations, fun, ..., protect_cues = TRUE)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR](ar_import).}

\item{fun}{a vectorized \code{function} applied to all responses and cues in the response matrix, excluding cues that exist in the .}

\item{...}{additional arguments passed to \code{fun}.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{p_id} to identify participants, and potentially additional participant attributes \code{p_attr_*}.}
 \item{cues}{A tibble of all cues \code{c_cue} and potentially additional cue attributes \code{c_attr_*}.}
 \item{responses}{A tibble of all participant's \code{p_id} cue \code{r_cue} and response \code{r_response} pairs including response levels \code{r_level}, as well as potentially additional responseattributes \code{r_attr_*}.}
}
}
\description{
\code{ar_normalize_manual} offers an interface to add self-defined normalization steps for responses (and cues) in associatoR objects.
}
\examples{

ai_asso <- ai_asso \%>\% ar_normalize_manual(tolower)
ai_asso <- ai_asso \%>\% ar_normalize_manual(trimws, which = "left")

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
