% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ar_plot_embedding}
\alias{ar_plot_embedding}
\title{Plot embedding}
\usage{
ar_plot_embedding(
  associations,
  facet_col = cluster,
  facet_row = NULL,
  color_by = cluster,
  top_n = 30,
  ...
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import} with targets defined by \link[associatoR]{ar_set_targets}.}

\item{facet_col}{name of variable to generate column facets by.}

\item{facet_row}{name of variable to generate row facets by.}

\item{color_by}{name of variable used to color the words.}

\item{top_n}{a \code{integer} value specifying the number of highest frequency words to display.}

\item{...}{}
}
\value{
Returns a \link[ggplot2]{ggplot} plot object.
}
\description{
Generate two dimensional plot for \code{associatoR} object.
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets("cues") \%>\%
  plot(facets = "gender", top_n = 10)

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets("cues") \%>\%
  plot(facets = "education", top_n = 10, gender == "female")

}
