% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{ar_project}
\alias{ar_project}
\title{Project embeddings}
\usage{
ar_project(associations, type = "umap", ...)
}
\arguments{
\item{associations}{an \code{ar_object} including targets.}

\item{type}{a \code{character} specifying the type of embedding. One of \code{c("counts","ppmi","ppmi-svd","huggingface")}. Default is \code{"ppmi-svd"}.}

\item{...}{additional parguments passed to the projection method.}
}
\value{
The function returns the \code{associatoR} object with the
  \code{target_embeddings} overwritten by the projected embeddings.
}
\description{
\code{ar_project} generates a 2D projection of the target embedding
}
\examples{

# run embedding
ar_project(ai_asso_processed)

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv
}
