% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_set_target}
\alias{ar_set_target}
\title{\code{ar_define_target} defines the target level to further analyze in an associatoR object. This could be multiple cue words to analyze based on responses, response words to analyze based on an external embedding, or response words to analyze based on a second level of responses.}
\usage{
ar_set_target(
  associations,
  source = c("cues", "responses"),
  response_level = 1
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR](ar_import).}

\item{source}{a \code{character} of \code{c("cues", "responses")} specifying, which column in Participants contains targets.}

\item{response_level}{an optional positive \code{integer} specifying, which response-level contains targets (defaults to 1).}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{p_id} to identify participants, and potentially additional participant attributes \code{p_attr_*}.}
 \item{cues}{A tibble of all cues \code{c_cue} and potentially additional cue attributes \code{c_attr_*}.}
 \item{responses}{A tibble of all participant's \code{p_id} cue \code{r_cue} and response \code{r_response} pairs including response levels \code{r_level}, as well as potentially additional responseattributes \code{r_attr_*}.}
 \item{targets}{A tibble of \code{target}s that will be analyzed by subsequent functions.}
}
}
\description{
\code{ar_define_target} defines the target level to further analyze in an associatoR object. This could be multiple cue words to analyze based on responses, response words to analyze based on an external embedding, or response words to analyze based on a second level of responses.
}
\examples{

ai_asso_imported <- ai_asso_imported \%>\%
  ar_define_target(source = "responses")

}
