% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_set_targets}
\alias{ar_set_targets}
\title{Set analysis targets}
\usage{
ar_set_targets(associations, targets = c("cues", "responses"))
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{targets}{a \code{character} containing either one or both of \code{c("cues","responses")} to gather the targets from these columns or a vector of manually selected targets.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
 \item{targets}{A tibble of targets including the specified analysis target}
}
}
\description{
\code{ar_set_target} defines the analysis targets (cues and/or responses).
}
\examples{

ai_asso <- ai_asso \%>\%
  ar_define_targets(source = "cues")

}
