% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_summarize_targets}
\alias{ar_summarize_targets}
\title{Add response variable summaries to targets}
\usage{
ar_summarize_targets(associations, response_var, fun = mean, ...)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{response_var}{a variable in the responses tibble of the \code{associatoR} object.}

\item{fun}{a function used to summarize response_vars for each target, defaults to \code{mean}.}

\item{...}{additional parameters passed on to \code{fun}.}
}
\value{
Returns an \code{associatoR} object where \code{targets} gains an additional column including the target summary calculated from the \code{responses} table's \code{response_var} using \code{fun}.
}
\description{
Add response variable summaries to targets
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_summarize_targets(response_position, fun = median)
}
