% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_target_add_frequency}
\alias{ar_target_add_frequency}
\title{Add a frequency variable to targets}
\usage{
ar_target_add_frequency(x, source = c("cues", "responses"))
}
\arguments{
\item{x}{an \code{associatoR} object generated by associatoR::ar_import()}

\item{source}{one of \code{c("cues", "responses")}, specifying where the targets come from.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{p_id} to identify participants, and potentially additional participant attributes \code{p_attr_*}.}
 \item{cues}{A tibble of all cues \code{c_cue} and potentially additional cue attributes \code{c_attr_*}.}
 \item{responses}{A tibble of all participant's \code{p_id} cue \code{r_cue} and response \code{r_response} pairs including response levels \code{r_level}, as well as potentially additional responseattributes \code{r_attr_*}.}
 \item{targets}{A tibble of \code{target} and \code{frequency} that will be analyzed by subsequent functions.}
}
}
\description{
Adds a variable \code{frequency} to the Targets in an associatoR object .
}
\examples{
ar_import(ai_asso, participant_id = "id",
                              response_col = "association_correct",
                              cue_type = "manual", cue_manual = "artificial intelligence",
                              participant_attributes = c("age", "gender", "use", "expertise"),
                              response_attributes = c("association", "trial")) \%>\%
  ar_normalize() \%>\%
  ar_define_target(source = "responses") \%>\%
  ar_target_add_frequency(source = "responses")

}
