% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ar_word_cloud}
\alias{ar_word_cloud}
\title{Plot word clouds}
\usage{
ar_word_cloud(
  associations,
  facet_col = cluster,
  facet_row = NULL,
  color_by = cluster,
  top_n = 30,
  ...
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import} with targets defined by \link[associatoR]{ar_set_targets}.}

\item{color_by}{a \code{character} vector of variable names to generate facets.}

\item{top_n}{a \code{numeric} scalar}

\item{...}{optional \code{logical} comparisons to filter the responses before computing frequencies.}

\item{facet_by}{names of variables to generate facets by.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
}
}
\description{
Generate word cloud plots of response frequencies for \code{associatoR} object.
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets("cues") \%>\%
  plot(facets = "gender", top_n = 10)

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets("cues") \%>\%
  plot(facets = "education", top_n = 10, gender == "female")

}
