% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{ar_wordlist_export}
\alias{ar_wordlist_export}
\title{Export wordlist}
\usage{
ar_wordlist_export(
  associations,
  file = "wordlist_correction.csv",
  check_spelling = TRUE,
  check_acronym = TRUE
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{file}{a \code{character} string specifying the filename including folders. Default is \code{"wordlist_correction.csv"}.}

\item{check_spelling}{a \code{logical} specifying whether to include a column identifying likely mispelled words based on the \link[spelling]{spell_check_text}. Note that currently the spell check only works for English associations. Default is \code{TRUE}.}

\item{check_acronyms}{a \code{character} specifying whether to include columns identifying potential acronyms. Acronyms are identified in two first. First, responses are checked against a Wikipedia list ("https://en.wikipedia.org/wiki/Lists_of_acronyms"). Matching of the original and lower-case responses triggers "likely an acronym" and "potentially an acronym". The matched acronym is spelled out in an additional column named \code{acronym_candidate}. Second, the a simple acronym detection based on letter casing also triggers "likely an acronym" but does not produce values in \code{acronym_candidate}. Default is \code{TRUE}.}
}
\value{
Writes a .CSV file containing a list of unique responses plus additional information.
}
\description{
\code{ar_wordlist_export()} exports a table of responses for manual spelling correction.
}
\examples{

ar_obj = ar_import(intelligence,
                   participant = participant_id,
                   cue = cue,
                   response = response,
                   participant_vars = c(gender, education),
                   response_vars = c(response_position, response_level)) \%>\%
  ar_normalize_manual(trimws, which = "left")

ar_wordlist_export(ar_obj)

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in preparation)
}
