% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{ar_wordlist_import}
\alias{ar_wordlist_import}
\title{Import wordlist}
\usage{
ar_wordlist_import(
  associations,
  file = "wordlist_correction.csv",
  process_cues = TRUE,
  na = c("NA")
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{file}{a \code{character} string specifying the filename containing the file spelling correction table exported by \link[associatoR]{ar_wordlist_export}. Must be a .CSV file and contain columns \code{c("response", "response_correct")}. Default name is \code{"wordlist_correction.csv"}.}

\item{process_cues}{a \code{logical} indicating if cues should be processed (i.e. changed by \code{fun}) or not.}

\item{na}{a \code{character} vector indicating the values signifying invalid values. Devaults to \code{c("NA")}.}
}
\value{
Writes a .CSV file containing a list of unique responses plus additional information.
}
\description{
\code{ar_wordlist_import()} import a table of responses from manual spelling correction.
}
\examples{

ar_obj = ar_import(intelligence,
                   participant = participant_id,
                   cue = cue,
                   response = response,
                   participant_vars = c(gender, education),
                   response_vars = c(response_position, response_level)) \%>\%
  ar_normalize_manual(trimws, which = "left")

ar_wordlist_import(ar_obj)

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in preparation)
}
