% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_data_counts.R
\name{making_data_counts}
\alias{making_data_counts}
\title{Making Data Counts}
\usage{
making_data_counts(data, target_cols, thresh = 5, title)
}
\arguments{
\item{data}{Dataset as a dataframe with the timepoint vairable as year column}

\item{target_cols}{The phenotype or column names for making the count}

\item{thresh}{Number of years as threshold, Has been default to 5}

\item{title}{The title of the phenotype for the graph}
}
\value{
Returns a barplot with counts for non-NA values inside target columns
}
\description{
Making Data Counts
}
\examples{
making_data_counts(data, target_cols, 'Phenotype 1')
}
