% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.associatoR}
\alias{plot.associatoR}
\title{Plot associations}
\usage{
\method{plot}{associatoR}(associations, facet_by = NULL, top_n = 5, ...)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import} with targets defined by \link[associatoR]{ar_set_targets}.}

\item{facet_by}{a \code{character} vector of variable names to generate facets.}

\item{top_n}{a \code{numeric} scalar}

\item{...}{optional \code{logical} comparisons to filter the responses before computing frequencies.}
}
\value{
Returns a \link[ggplot2]{ggplot} plot object.
}
\description{
Generate a plot of an \code{associatoR} object.
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets("cues") \%>\%
  plot(facets = "gender", top_n = 10)

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets("cues") \%>\%
  plot(facets = "education", top_n = 10, gender == "female")

}
