% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.associatoR}
\alias{print.associatoR}
\title{Generic print for class associatoR}
\usage{
\method{print}{associatoR}(x, n = 5, ...)
}
\arguments{
\item{x}{an object of class \code{associatoR}}

\item{n}{(optional) use \code{print(n = )} to show more rows of each tibble.}

\item{...}{additional arguments passed to \code{print()}}
}
\description{
Generic print for class associatoR
}
