% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_sectional_association.R
\name{run_cross_sectional_association}
\alias{run_cross_sectional_association}
\title{Cross Sectional Association}
\usage{
run_cross_sectional_association(
  data,
  targets,
  timepoint,
  formula,
  fdr_correction = FALSE
)
}
\arguments{
\item{data}{The dataframe with all covariates, targets and elements included in formula after having transformed the data}

\item{targets}{The dependant variable (phenotype) that is to be associated with}

\item{timepoint}{The timepoint which is to be analyzed and can be personlized to the timepoint variable}

\item{formula}{formula for providing the association without the dependant variable. Look at example for how to define the formula}

\item{fdr_correction}{Default to False but if FDR correction is required, the p-values are adjusted for the number of features}
}
\value{
Associations with the p-values, coefficients and variance explained for all the features for each phenotype or target
}
\description{
Cross Sectional Association
}
