% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_sectional_timepoints.R
\name{run_cross_sectional_timepoints}
\alias{run_cross_sectional_timepoints}
\title{Running cross-sectional analysis for all timepoints}
\usage{
run_cross_sectional_timepoints(
  data,
  targets,
  timepoints,
  formula,
  fdr_correction = FALSE
)
}
\arguments{
\item{data}{The dataframe with all covariates, targets and elements included in formula after having transformed the data}

\item{targets}{The dependant variable (phenotype) that is to be associated with}

\item{timepoints}{The timepoints which are to be analyzed and can be personlized to the timepoint variable}

\item{formula}{formula for providing the association without the dependant variable. Look at example for how to define the formula}

\item{fdr_correction}{Default to False but if FDR correction is required, the p-values are adjusted for the number of features}
}
\value{
Associations with the p-values, coefficients and variance explained for each timepoint and for all the features for each phenotype or target
}
\description{
Running cross-sectional analysis for all timepoints
}
