% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_longitudinal_interaction_association.R
\name{run_longitudinal_interaction_association}
\alias{run_longitudinal_interaction_association}
\title{Running linear mixed models for all timepoints}
\usage{
run_longitudinal_interaction_association(
  data,
  targets,
  threshold_years = 5,
  formula,
  plot_effect = TRUE,
  plot_residuals = FALSE
)
}
\arguments{
\item{data}{The dataframe with all covariates, targets and elements included in formula after having transformed the data}

\item{targets}{The dependant variable (phenotype) that is to be associated with}

\item{threshold_years}{Number of years as threshold, Has been default to 5}

\item{formula}{formula for providing the association without the dependant variable. Look at example for how to define the formula}

\item{plot_effect}{Default to True. Provides effect plots for the linear mixed models fit}

\item{plot_residuals}{Default to False Provides residual plots for the linear mixed models fit}
}
\value{
list of one dataframe that provides the p-value of the phenotype or target, and if plot_effect is TRUE, returns effect plot for each phenotype
    and if plot_residuals is TRUE, return residula plots for the fit
}
\description{
Running linear mixed models for all timepoints
}
