% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_characterize_targets}
\alias{ar_characterize_targets}
\title{Characterize targets using additional resources}
\usage{
ar_characterize_targets(
  associations,
  characteristics = c("valence", "arousal", "dominance", "concreteness", "frequency")
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{characteristics}{a \code{character} string or vector specifying the characteristics using one or more from \code{c("valence", "arousal", "dominance", "concreteness", "frequency")} or a dataset containing characterisics. The dataset must contain a column \code{word}.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles, with \code{targets} gaining additional columns including the target characteristics:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
 \item{targets}{A tibble of targets including the specified analysis target, and word characteristics}
}
}
\description{
Append target characteristic to the \code{targets} table in the \code{associatoR} object. Specify characteristics and or manual look up table.
}
\details{
Data on valence, arousal, and dominance are retrieved from Warriner et al. (2013), concreteness and frequency data are retrieved from Brysbaert et al. (2014).
}
\examples{
ar_import(fa_data,
          participant = participantID, participant_vars = c(age, gender),
          cue = cue, response = response,
          response_vars = c("created_at", "pos")) \%>\%
  ar_normalize() \%>\%
  ar_set_targets("cues") \%>\%
  ar_characterize_targets()

}
\references{
\itemize{
  \item{Warriner, A. B., Kuperman, V., & Brysbaert, M. (2013). Norms of valence, arousal, and dominance for 13,915 English lemmas. Behavior Research Methods, 45(4), 1191–1207. https://doi.org/10.3758/s13428-012-0314-x}
  \item{Brysbaert, M., Warriner, A. B., & Kuperman, V. (2014). Concreteness ratings for 40 thousand generally known English word lemmas. Behavior Research Methods, 46(3), 904–911. https://doi.org/10.3758/s13428-013-0403-5}
}
}
