% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_characterize_targets}
\alias{ar_characterize_targets}
\title{Characterize targets using additional resources}
\usage{
ar_characterize_targets(
  associations,
  characteristics = c("valence", "arousal", "dominance", "concreteness",
    "word_frequency")
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{characteristics}{a \code{character} string or vector specifying the characteristic(s) containing one or more from \code{c("valence", "arousal", "dominance", "concreteness", "word_frequency")} or a \code{data.frame} containing characteristics. The \code{data.frame} must contain a column \code{word} to serve as the look up key and may contain additional columns with word characteristics.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles, with \code{targets} gaining additional columns including the target characteristics:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
 \item{targets}{A tibble of targets including the specified analysis targets, and word characteristics}
}
}
\description{
Append target characteristics to the \code{targets} table in the \code{associatoR} object. Use pre-defined characteristics from psycho-linguistic research or supply a data.frame as a user-defined look-up table.
}
\details{
Data on valence, arousal, and dominance are retrieved from Warriner et al. (2013). Concreteness and word frequency data are retrieved from Brysbaert et al. (2014).
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_characterize_targets()

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_characterize_targets(characteristics = data.frame(word = c("intelligence", "brain"), message = c("Hello", "World")))

}
\references{
\itemize{
  \item{Warriner, A. B., Kuperman, V., & Brysbaert, M. (2013). Norms of valence, arousal, and dominance for 13,915 English lemmas. Behavior Research Methods, 45(4), 1191–1207. https://doi.org/10.3758/s13428-012-0314-x}
  \item{Brysbaert, M., Warriner, A. B., & Kuperman, V. (2014). Concreteness ratings for 40 thousand generally known English word lemmas. Behavior Research Methods, 46(3), 904–911. https://doi.org/10.3758/s13428-013-0403-5}
}
}
