% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{ar_compare}
\alias{ar_compare}
\title{Compare responses}
\usage{
ar_compare(
  associations,
  participant_vars,
  target_var = "targets",
  fun = c("mean", "count"),
  normalize = NULL
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{participant_vars}{a \code{character} vector specifying the grouping variables.}

\item{target_var}{an optional \code{character} string specifying the target variable to be summarized. If not specified \code{ar_comapre} tabulates the frequencies of targets.}

\item{normalize}{an optional \code{character} string specifying whether the output should be normalized by rows (\code{normalize = "rows"}) or columns (\code{normalize = "columns"}). Defaults to \code{normalize = NULL} implying no normalization.}
}
\value{
Returns an \code{tibble} conatining grouped summaries (means or frequencies).
}
\description{
\code{ar_compare} compares responses across participant variables.
}
\examples{

ai_asso <- ai_asso \%>\%
  ar_define_targets(source = "cues")

}
