% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{ar_compare}
\alias{ar_compare}
\title{Compare responses}
\usage{
ar_compare(
  associations,
  participant_vars,
  target_var,
  fun = c("count", "mean")
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{participant_vars}{a \code{character} vector specifying the grouping variables.}

\item{target_var}{an optional \code{character} string specifying the target variable to be summarized. If not specified \code{ar_comapre} tabulates the frequencies of targets.}

\item{fun}{a \code{character} string or a function specifying the statistic to calculate from \code{target_var}.}
}
\value{
Returns an \code{tibble} containing grouped summaries (means, frequencies, or other custom statistics).
}
\description{
\code{ar_compare} compares responses across participant variables.
}
\examples{

ai_asso \%>\%
  ar_ar_compare()

}
