% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{ar_cross_targets}
\alias{ar_cross_targets}
\title{Create target participant cross tables}
\usage{
ar_cross_targets(associations, participant_vars, target_var, normalize = FALSE)
}
\arguments{
\item{normalize}{}
}
\description{
Create target participant cross tables
}
