% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_define_target}
\alias{ar_define_target}
\title{Define target in an associatoR object}
\usage{
ar_define_target(x, source, level = NULL)
}
\arguments{
\item{x}{an \code{associatoR} object generated by associatoR::ar_import()}

\item{level}{an \code{integer} (otional) specifying, which response-level contains targets.}

\item{souce}{one of \code{c("cue", "response")} specifying, which column in $participants contains targets.}
}
\value{
Returns an \code{associatoR} object containing a list of \code{tibble}s:
\describe{
 \item{participants}{A tibble of all participants including \code{part_id}, and potentially additional participant-specific attributes \code{part_*}.}
 \item{cues}{A tibble of all cues including \code{cue_id}, \code{cue_text} and potentially additional cue-specific attributes \code{cue_*}.}
 \item{responses}{A tibble of all responses including \code{resp_id}, \code{resp_text} and potentially additional response-specific attributes \code{resp_*}.}
 \item{targets}{A tibble of targets to be further analyzed including \code{target_id}, \code{target_text}, and }
}
}
\description{
\code{ar_define_target} defines the target level to further analyze in an associatoR object. This could be multiple cue words to analyze based on responses, response words to analyze based on an external embedding, or response words to analyze based on a second level of responses.
}
\examples{

ai_asso <- ar_define_target(ai_asso)
}
