% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import}
\alias{ar_import}
\title{Import association data into associatoR object}
\usage{
ar_import(
  data,
  part_id,
  part_attr = c(),
  cue_type = c("single", "multiple"),
  cue_single = NULL,
  cue_text = NULL,
  cue_attr = c(),
  resp_text,
  resp_attr = c()
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble} containing the association data.}

\item{part_id}{a \code{character} scalar of the variable name in \code{data} identifying individual participants.}

\item{part_attr}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of participants.}

\item{cue_type}{one of \code{c("single", "multiple")}; if \code{cue_type = "single"}, the parameter \code{cue_single} must be specified, if \code{cue_type = "multiple"}, the parameter \code{cue_id} must be specified.}

\item{cue_single}{a \code{character} scalar to specify a cue that was used for all responses.}

\item{cue_text}{a \code{character} scalar of the variable name in \code{data} identifying individual cues.}

\item{cue_attr}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of cues.}

\item{resp_text}{a \code{character} scalar of the variable name in \code{data} identifying individual responses.}

\item{resp_attr}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of responses.}
}
\value{
Returns an \code{associatoR} object containing a list of \code{tibble}s:
\describe{
 \item{participants}{A tibble of all participants including \code{part_id}, and potentially additional participant-specific attributes \code{part_*}.}
 \item{cues}{A tibble of all cues including \code{cue_id}, \code{cue_text} and potentially additional cue-specific attributes \code{cue_*}.}
 \item{responses}{A tibble of all responses including \code{resp_id}, \code{resp_text} and potentially additional response-specific attributes \code{resp_*}.}
}
}
\description{
\code{ar_import} processes tabular association data into an associatoR object
}
\examples{

# import association data object
ar_obj <- ar_import(data = ai_asso, part_id = "id",
                    part_attr = c("age", "gender", "use", "expertise"),
                    cue_type = "single", cue_single = "AI",
                    resp_text = "association_correct",
                    resp_attr = c("association", "trial"))

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress). associatoR. psyArXiv
}
