% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import}
\alias{ar_import}
\title{Import R data into associatoR object}
\usage{
ar_import(data)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble} object containing the association data.}

\item{participant_labels}{a \code{character} vector ... .}
}
\value{
The function returns the \code{associatoR} object, which contains a list with the following elements, each a \code{tibble}:
\describe{
 \item{participant}{Index of the element in \code{text} where match was found. Formatted as a factor with the number of levels matching the original number of documents.}
 \item{cues}{Label of the SDG found in document.}
 \item{responses}{The name of the ensemble system that produced the match.}
 \item{hit}{Index of hit for the Ensemble model.}
}
}
\description{
\code{ar_import} processes R data and generates an associatoR object
}
\examples{

# run sdg detection
data <- ar_impot(ai_asso)

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv
}
