% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import}
\alias{ar_import}
\title{Import association data into associatoR object}
\usage{
ar_import(
  data,
  part_id,
  part_attr = c(),
  cue_type = c("single", "multiple"),
  cue_single = NULL,
  cue_text = NULL,
  cue_attr = c(),
  resp_text,
  resp_attr = c()
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble} containing the association data.}

\item{part_id}{a \code{character} scalar of the variable name in \code{data} identifying individual participants.}

\item{part_attr}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of participants.}

\item{cue_type}{one of \code{c("single", "multiple")}; if \code{cue_type = "single"}, the parameter \code{cue_single} must be specified, if \code{cue_type = "multiple"}, the parameter \code{cue_id} must be specified.}

\item{cue_single}{a \code{character} scalar to specify a cue that was used for all responses.}

\item{cue_text}{a \code{character} scalar of the variable name in \code{data} identifying individual cues.}

\item{cue_attr}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of cues.}

\item{resp_text}{a \code{character} scalar of the variable name in \code{data} identifying individual responses.}

\item{resp_attr}{a \code{character} vector containing all additional variable names in \code{data} that are attributes of responses.}
}
\value{
Returns an \code{associatoR} object containing a list of \code{tibble}s:
\describe{
 \item{participants}{}
 \item{cues}{}
 \item{responses}{}
 \item{targets}{}
}
}
\description{
\code{ar_import} processes tabular association data into an associatoR object
}
\examples{

# import association data object
ar_obj <- ar_import(ai_asso)

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv
}
