% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import}
\alias{ar_import}
\title{Import association data into an associatoR object}
\usage{
ar_import(
  data,
  participant,
  cue = NULL,
  cue_manual = NULL,
  response,
  response_level = NULL,
  participant_vars = NULL,
  cue_vars = NULL,
  response_vars = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble} containing the association data.}

\item{participant}{a \code{character} scalar of the variable name in \code{data} identifying participants.}

\item{cue}{a \code{character} scalar of the variable name in \code{data} identifying the cues.}

\item{cue_manual}{an optional \code{character} scalar specifying the cue for all responses. Overrides argument \code{cue}.}

\item{response}{a \code{character} scalar of the variable name in \code{data} identifying responses.}

\item{response_level}{an optional \code{character} scalar of the variable name in \code{data} identifying response levels.}

\item{participant_vars}{an optional \code{character} vector containing all additional variable names in \code{data} that are attributes of participants.}

\item{cue_vars}{an optional \code{character} vector containing all additional variable names in \code{data} that are attributes of cues.}

\item{response_vars}{an optional \code{character} vector containing all additional variable names in \code{data} that are attributes of responses}

\item{verbose}{an \code{logical} specifying whether to show messages.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{p_id} to identify participants, and potentially additional participant attributes \code{p_attr_*}.}
 \item{cues}{A tibble of all cues \code{c_cue} and potentially additional cue attributes \code{c_attr_*}.}
 \item{responses}{A tibble of all participant's \code{p_id} cue \code{r_cue} and response \code{r_response} pairs including response levels \code{r_level}, as well as potentially additional responseattributes \code{r_attr_*}.}
}
}
\description{
\code{ar_import} processes tabular association data into an associatoR object.
}
\examples{
ai_asso_imported <- ar_import(ai_asso,
                              participant = "id",
                              response = "association_correct",
                              cue_manual = "AI",
                              participant_vars = c("age", "gender", "use", "expertise"),
                              response_vars = c("association", "trial"))

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
