% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ar_import}
\alias{ar_import}
\title{Title}
\usage{
ar_import(
  data,
  participant_id,
  response_col,
  cue_type,
  cue_col = NULL,
  cue_manual = NULL,
  response_level = NULL,
  participant_attributes = NULL,
  cue_attributes = NULL,
  response_attributes = NULL
)
}
\arguments{
\item{response_attributes}{}
}
\description{
Title
}
