% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize}
\alias{ar_normalize}
\title{Normalize words in an associatoR object}
\usage{
ar_normalize(
  x,
  case = "lower",
  punct = "end",
  whitespace = "squish",
  normalize_cues = FALSE
)
}
\arguments{
\item{x}{an \code{associatoR} object generated by \code{associatoR::ar_import()}}

\item{case}{one of \code{c("lower", "upper", "original")}, converts the responses (and cues) to the specified casing, \code{"original"} does not change the casing.}

\item{punct}{one of \code{c("end", "all", "original")}, \code{end} replaces all punctuation at the end (or only followed by whitespace) of the string with whitespace, \code{"all"} replaces all punctuation characters in the string by whitespace, \code{"original"} changes nothing.}

\item{whitespace}{one of \code{c("squish", "trim", "original")}, \code{"squish"} removes whitespace from the ends of the string, an reduces multiple whitespace between characters to one, \code{"trim"} removes whitespace from the ends of the string, \code{"original"} keeps whitespaces as is (note that additional whitespaces can be generated based on the string passed to the \code{punct} argument).}

\item{normalize_cues}{(\code{FALSE}) change to \code{TRUE} to apply normalization steps to cues as well.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{p_id} to identify participants, and potentially additional participant attributes \code{p_attr_*}.}
 \item{cues}{A tibble of all cues \code{c_cue} and potentially additional cue attributes \code{c_attr_*}.}
 \item{responses}{A tibble of all participant's \code{p_id} cue \code{r_cue} and response \code{r_response} pairs including response levels \code{r_level}, as well as potentially additional responseattributes \code{r_attr_*}.}
}
}
\description{
\code{ar_normalize} inlcludes often used normalization steps for responses (and cues) and wraps around \code{ar_normalize_manual}.
}
\examples{

ai_asso_processed <- ar_normalize(ai_asso_imported)
ai_asso_processed <- ar_normalize(ai_asso_imported, case = "original",
                                  punct = "all")

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
