% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize}
\alias{ar_normalize}
\title{Normalize responses}
\usage{
ar_normalize(
  associations,
  case = "lower",
  punct = "end",
  punct_replacement = " ",
  whitespace = "squish",
  process_cues = FALSE
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{case}{a \code{character} specifying the normalization of cases. Must be one of \code{c("lower", "upper", "sentence")}. Setting \code{case = "lower"} normalizes all responses to lower case, \code{case = "lower"} to upper case, and \code{case = "sentence"} to sentence case.}

\item{punct}{a \code{character} specifying the normalization of punctuation. Must be one of \code{c("end", "all")}. \code{punct = "end"} replaces all punctuation at the end (including following whitespaces) with \code{punct_replacement} (default is a single white space). \code{punct = "all"} replaces all punctuation.}

\item{punct_replacement}{a \code{character} used as replacement for punctuation.}

\item{whitespace}{a \code{character} specifying the normalization of white spaces. Must be one of \code{c("squish", "trim")}. Setting \code{whitespace = "squish"} removes additional white spaces at the start, the end, and in-between words, whereas setting \code{whitespace = "trim"} removes white space from the start and end.}

\item{process_cues}{a \code{logical} specifying, if cues should be processed as well. Defaults to \code{FALSE}.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
}
}
\description{
\code{ar_normalize} performs normalization steps for responses (and cues) and wraps around \link[associatoR]{ar_normalize_manual}.
}
\examples{

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_normalize(case = "lower", punct = "all")

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
