% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize_manual}
\alias{ar_normalize_manual}
\title{Normalize responses manually}
\usage{
ar_normalize_manual(associations, fun, ..., process_cues = FALSE)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{fun}{a vectorized \code{function} applied to all responses and cues in the response matrix, excluding cues that exist in the .}

\item{...}{additional arguments passed to \code{fun}.}

\item{process_cues}{a \code{logical} indicating if cues should be processed (i.e. changed by \code{fun}) or not.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:

\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes. All responses are passed through \code{fun}}
}
}
\description{
\code{ar_normalize_manual} offers an interface to add self-defined normalization steps for responses (and cues) in \code{associatoR} objects.
}
\examples{

ar_import(ai_asso, participant = id, cue = "AI",
          response = association_correct,
          participant_vars = c(age, gender, use, expertise),
          response_vars = c(trial)) \%>\%
  ar_normalize_manual(trimws, which = "left")

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
