% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize_manual}
\alias{ar_normalize_manual}
\title{Normalize words in an associatoR object manually}
\usage{
ar_normalize_manual(x, fun, normalize_cues = FALSE, ...)
}
\arguments{
\item{x}{an \code{associatoR} object generated by associatoR::ar_import()}

\item{fun}{a \code{function} to apply to all responses (resp_text) in an associatoR object.}

\item{normalize_cues}{(\code{FALSE}) change to \code{TRUE} to apply \code{fun} to cues as well.}

\item{...}{additional arguments passed on to \code{fun}.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{p_id} to identify participants, and potentially additional participant attributes \code{p_attr_*}.}
 \item{cues}{A tibble of all cues \code{c_cue} and potentially additional cue attributes \code{c_attr_*}.}
 \item{responses}{A tibble of all participant's \code{p_id} cue \code{r_cue} and response \code{r_response} pairs including response levels \code{r_level}, as well as potentially additional responseattributes \code{r_attr_*}.}
}
}
\description{
\code{ar_normalize_manual} offers an interface to add self-defined normalization steps for responses (and cues) in associatoR objects.
}
\examples{

ai_asso_processed <- ar_normalize_manual(ai_asso_imported, tolower)
ai_asso_processed <- ar_normalize_manual(ai_asso_imported, trimws, which = "left")

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress)
}
