% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ar_normalize_manual}
\alias{ar_normalize_manual}
\title{Normalize responses in an associatoR object manually}
\usage{
ar_normalize_manual(x, fun, ...)
}
\arguments{
\item{x}{an \code{associatoR} object generated by associatoR::ar_import()}

\item{fun}{a \code{function} to apply to all responses (resp_text) in an associatoR object.}

\item{...}{additional arguments passed on to \code{fun}.}

\item{case}{one of \code{c("lower", "upper", "original"), converts the resp_text to the specified casing, \code{"original"} does not change the casing.}}
}
\value{
Returns an \code{associatoR} object containing a list of \code{tibble}s with \code{fun} applied to all responses:
\describe{
 \item{participants}{A tibble of all participants including \code{part_id}, and potentially additional participant-specific attributes \code{part_*}.}
 \item{cues}{A tibble of all cues including \code{cue_id}, \code{cue_text} and potentially additional cue-specific attributes \code{cue_*}.}
 \item{responses}{A tibble of all responses including \code{resp_id}, \code{resp_text} and potentially additional response-specific attributes \code{resp_*}.}
}
}
\description{
\code{ar_normalize_manual(x, fun, ...)} offers an interface to add self-defined normalization steps for responses in associatoR objects.
}
\examples{

ai_asso_processed <- ar_normalize_manual(ai_asso_processed, tolower)

ai_asso_processed <- ar_normalize_manual(ai_asso_processed, trimws, which = "left")

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in progress). associatoR. psyArXiv
}
