% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ar_plot_embedding}
\alias{ar_plot_embedding}
\title{Plot embedding}
\usage{
ar_plot_embedding(
  associations,
  color_by = NULL,
  color_set = "G",
  alpha = 0.5,
  proportion_labels = 1,
  ...
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import} with targets defined by \link[associatoR]{ar_set_targets}.}

\item{color_by}{name of variable to generate column facets by.}

\item{color_set}{a \code{character} string specifying the color set. One of \code{c("A","B","C","D","E","F","G","H")}. See \link[ggplot2]{scale_color_viridis_d}.}

\item{alpha}{a \code{numeric} value between 0 and 1 specifying the transparency of points. Default is \code{.5}.}

\item{proportion_labels}{a \code{numeric} value between 0 and 1 specifying the proportion of labels shown in the plot. Values smaller than 1 omit 1-\code{proportion_labels} of labels for lower frequency targets.}

\item{...}{optional arguments passed on to \link[ggrepel]{geom_text_repel}.}
}
\value{
Returns a \link[ggplot2]{ggplot} plot object.
}
\description{
Generate two dimensional plot for \code{associatoR} object.
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets("cues") \%>\%
  ar_embed_targets() \%>\%
  ar_cluster_targets() \%>\%
  ar_project_targets() \%>\%
  ar_plot_embedding(color_by = cluster,
                    proportion_labels = .5)

}
