% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{ar_project}
\alias{ar_project}
\title{Project embeddings}
\usage{
ar_project(associations, type = "umap", ...)
}
\arguments{
\item{associations}{an \code{associatoR} object including targets.}

\item{type}{a \code{character} specifying the type of projection. One of \code{c("mds", "umap")}. Default is \code{"umap"}.}

\item{...}{additional parguments passed to the projection method.}
}
\value{
The function returns the \code{associatoR} object with the
  \code{target_embeddings} overwritten by the projected embeddings.
}
\description{
\code{ar_project} generates a 2D projection of the target embedding.
}
\examples{
ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_set_targets(targets = "cues") \%>\%
  ar_embed() \%>\%
  ar_project()

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv
}
