% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_set_targets}
\alias{ar_set_targets}
\title{Set analysis targets}
\usage{
ar_set_targets(
  associations,
  target_set = c("cues", "responses"),
  target_manual = NULL
)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{target_manual}{a \code{character} character specifying the analysis target. Overrides argument \code{target_sets}.}

\item{target_sets}{a \code{character} string or vector specifying whether cues and/or responses should the targets of analysis. Must be one or both of \code{c("cues", "responses")}.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
 \item{targets}{A tibble of targets including the specified analysis target}
}
}
\description{
\code{ar_set_target} defines the analysis targets (cues and/or responses).
}
\examples{

ai_asso <- ai_asso \%>\%
  ar_define_targets(source = "cues")

}
