% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{ar_subset}
\alias{ar_subset}
\title{Subset associations}
\usage{
ar_subset(associations, ...)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{...}{optional \code{logical} comparisons filtering the responses before computing the frequencies.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
}
}
\description{
Subset the \code{associatoR} object.
}
\details{
Note that subsets resets the \code{associatoR} object. A potential \code{target} tibble will be removed and has to be added again using \link[associatoR]{ar_import}.
}
\examples{

ai_asso <- ai_asso \%>\%
  ar_subset(id > 10, trial = 10)

}
