% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{ar_subset}
\alias{ar_subset}
\title{Subset associations}
\usage{
ar_subset(associations, ...)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{...}{optional \code{logical} comparisons filtering the responses before computing the frequencies.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
}
}
\description{
Subset the \code{associatoR} object.
}
\details{
Note that subsets resets the \code{associatoR} object. A potential \code{target} tibble will be removed and has to be added again using \link[associatoR]{ar_set_target}.
}
\examples{

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) \%>\%
  ar_subset(gender == "female", education == "high school")

}
