% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{ar_target_add_frequency}
\alias{ar_target_add_frequency}
\title{Add a frequency variable to targets}
\usage{
ar_target_add_frequency(associations, source = c("cues", "responses"))
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR](ar_import).}

\item{source}{a \code{character} of \code{c("cues", "responses")} specifying, which column in Participants contains targets.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of all participants including \code{id} to identify participants, and potentially additional participant varables.}
 \item{cues}{A tibble of all cues \code{cue} and potentially additional cue variables.}
 \item{responses}{A tibble of all participant's \code{id} cue \code{r_cue} and response \code{response} pairs including response levels \code{level}, as well as potentially additional response variables}
 \item{targets}{A tibble of \code{target} and \code{frequency} that will be analyzed by subsequent functions.}
}
}
\description{
Adds a variable \code{frequency} to the Targets in an associatoR object .
}
\examples{
ar_import(ai_asso, participant_id = "id",
                              response_col = "association_correct",
                              cue_type = "manual", cue_manual = "artificial intelligence",
                              participant_attributes = c("age", "gender", "use", "expertise"),
                              response_attributes = c("association", "trial")) \%>\%
  ar_normalize() \%>\%
  ar_define_target(source = "responses") \%>\%
  ar_target_add_frequency(source = "responses")

}
