% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{ar_wordlist_export}
\alias{ar_wordlist_export}
\title{Export wordlist}
\usage{
ar_wordlist_export(associations, folder, file = "wordlist_correction.csv")
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{folder}{a \code{character} string specifying where to write the table.}

\item{file}{a \code{character} string specifying the filename. Default is \code{"wordlist.csv"}.}
}
\value{
Writes a .CSV file to the hard drive.
}
\description{
\code{ar_wordlist_export()} exports a table of responses for manual spelling correction.
}
\examples{

ar_obj = ar_import(intelligence,
                   participant = participant_id,
                   cue = cue,
                   response = response,
                   participant_vars = c(gender, education),
                   response_vars = c(response_position, response_level)) \%>\%
  ar_normalize_manual(trimws, which = "left")

ar_wordlist_export(ar_obj)

}
\references{
Aeschbach, S., Mata, R., Wulff, D. U. (in preparation)
}
