% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ar_write_csv}
\alias{ar_write_csv}
\title{Write a csv file}
\usage{
ar_write_csv(associations, file, ...)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{file}{a \code{character} to write to.}

\item{...}{optional arguments passed to readr::write_csv()}
}
\value{
\code{ar_write_csv()} returns the joined associations as a tibble invisibly.
}
\description{
\code{ar_write_csv()} joins all data in an \code{associatoR} object (except potential target embedding) into one table and writes the table to disk using readr::write_csv().
}
\examples{

ar_import(risk_asso, participant = id, cue = cue, response = response,
          response_vars = trial, participant_vars = c(gender, age, age_group)) \%>\%
 ar_normalize() \%>\%
 ar_set_targets(targets = "responses") \%>\%
 ar_write_csv(file = "associations.csv")

}
