% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.associatoR}
\alias{plot.associatoR}
\title{Plot associations}
\usage{
\method{plot}{associatoR}(associations, facets = NULL, top_n = 5, ...)
}
\arguments{
\item{associations}{an \code{associatoR} object containing association data as generated by \link[associatoR]{ar_import}.}

\item{...}{optional \code{logical} comparisons filtering the responses before computing the frequencies.}
}
\value{
Returns an \code{associatoR} object containing a list of tibbles:
\describe{
 \item{participants}{A tibble of participants including a participant \code{id} and potential participant attributes.}
 \item{cues}{A tibble of cues including a \code{cue} variable and potential cue attributes.}
 \item{responses}{A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.}
}
}
\description{
Plot the \code{associatoR} object.
}
